# api/openapi.yaml
openapi: 3.0.3
info:
  title: CompAInion API
  version: 0.1.0
  description: |
    REST backend for CompAInion: a Hermes session acting as "companion"
    dispatches plans/tasks to detached Hermes workers and serves the user one
    interruption at a time.

    Conventions:
    - Times are RFC3339 UTC strings.
    - JSON-valued columns (`handle`, `scope`, `acceptance_checks`, `evidence`,
      `payload`, `outcome`) are typed as their natural JSON shape here.
    - Optimistic concurrency: plans/steps/tasks carry `version`; PATCH takes
      `If-Match: "<version>"` (missing -> 428, stale -> 412). Responses carry an
      `ETag: "<version>"` header.
    - Every creating `POST` accepts `Idempotency-Key`; replay returns the cached
      response with `Idempotent-Replayed: true`, same key + different body -> 422.
    - Lists are cursor paginated: `?cursor=<last id>&limit=<n>` (default 50, max
      200), envelope `{items, next_cursor}`.
    - Errors are RFC 7807 `application/problem+json` (400/404/409/412/422/428).
    - Step/task `status` is the stored status; `ready` is a derived boolean
      (pending + all deps done). A `ready` value is exposed ONLY via this flag.
servers:
  - url: http://127.0.0.1:7777
...
paths:
  /workstreams:
    get:
      operationId: listWorkstreams
      tags: [workstreams]
      summary: List workstreams
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Paginated workstreams
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Page'
                  - type: object
                    required: [items]
                    properties:
                      items:
                        type: array
                        items: { $ref: '#/components/schemas/Workstream' }
        default: { $ref: '#/components/responses/Problem' }
    post:
      operationId: createWorkstream
      tags: [workstreams]
      summary: Create a workstream
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/WorkstreamInput' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Workstream' }
        '400': { $ref: '#/components/responses/Problem' }
        default: { $ref: '#/components/responses/Problem' }
  /workstreams/{id}:
    parameters:
      - $ref: '#/components/parameters/Id'
    get:
      operationId: getWorkstream
      tags: [workstreams]
      responses:
        '200':
          description: The workstream
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Workstream' }
        '404': { $ref: '#/components/responses/Problem' }
        default: { $ref: '#/components/responses/Problem' }
    patch:
      operationId: updateWorkstream
      tags: [workstreams]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/WorkstreamInput' }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Workstream' }
        '404': { $ref: '#/components/responses/Problem' }
        default: { $ref: '#/components/responses/Problem' }
    delete:
      operationId: deleteWorkstream
      tags: [workstreams]
      description: Refused (409) while plans or tasks still reference it.
      responses:
        '204': { description: Deleted }
        '409': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
        default: { $ref: '#/components/responses/Problem' }
...
  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
        - { name: workstream_id, in: query, schema: { type: string } }
        - { name: status, in: query, schema: { $ref: '#/components/schemas/WorkStatus' } }
      responses:
        '200':
          description: Paginated tasks
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TaskPage' }
        default: { $ref: '#/components/responses/Problem' }
    post:
      operationId: createTask
      tags: [tasks]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TaskInput' }
      responses:
        '201':
          description: Created (with `ETag`)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Task' }
        default: { $ref: '#/components/responses/Problem' }
  /tasks/{id}:
    parameters:
      - $ref: '#/components/parameters/Id'
    get:
      operationId: getTask
      tags: [tasks]
      responses:
        '200':
          description: The task (with `ETag`)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Task' }
        '404': { $ref: '#/components/responses/Problem' }
        default: { $ref: '#/components/responses/Problem' }
    patch:
      operationId: updateTask
      tags: [tasks]
      description: Only while `pending`. Requires `If-Match`.
      parameters:
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TaskInput' }
      responses:
        '200':
          description: Updated (with `ETag`)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Task' }
        '404': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
        '412': { $ref: '#/components/responses/Problem' }
        default: { $ref: '#/components/responses/Problem' }
    delete:
      operationId: deleteTask
      tags: [tasks]
      description: Only `pending`.
      responses:
        '204': { description: Deleted }
        '404': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
        default: { $ref: '#/components/responses/Problem' }
  /tasks/{id}/claim:
    parameters:
      - $ref: '#/components/parameters/Id'
    post:
      operationId: claimTask
      tags: [tasks]
      description: Same rules as step claim.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [agent_id]
              properties:
                agent_id: { type: string }
      responses:
        '200':
          description: Claimed
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Task' }
        '404': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
        default: { $ref: '#/components/responses/Problem' }
  /tasks/{id}/finish:
    parameters:
      - $ref: '#/components/parameters/Id'
    post:
      operationId: finishTask
      tags: [tasks]
      description: Same rules as step finish (`done` requires `outcome.evidence`).
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status, outcome]
              properties:
                status:
                  type: string
                  enum: [done, failed, blocked, interrupted]
                outcome: { $ref: '#/components/schemas/Outcome' }
      responses:
        '200':
          description: Finished
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Task' }
        '404': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
        '422': { $ref: '#/components/responses/Problem' }
        default: { $ref: '#/components/responses/Problem' }
...
components:
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: string }
    Cursor:
      name: cursor
      in: query
      description: Opaque cursor (the last id of the previous page).
      schema: { type: string }
    Limit:
      name: limit
      in: query
      description: Page size (default 50, max 200 for resources; 100/500 for events).
      schema: { type: integer, minimum: 1, maximum: 500, default: 50 }
    IfMatch:
      name: If-Match
      in: header
      description: 'Expected `version` as an ETag, e.g. "3". Required on PATCH'
      schema: { type: string }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      description: Client-supplied key caching the response of this creating
        POST. Same key + different body -> 422.
      schema: { type: string }
  responses:
    Problem:
      description: RFC 7807 problem details
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
    Problem:
      type: object
      description: RFC 7807 problem details. `type` maps the domain error kind
        (not_found, conflict, invalid, unprocessable, precondition_failed,
        precondition_required).
      required: [type, title, status]
      properties:
        type: { type: string, example: not_found }
        title: { type: string, example: Resource not found }
        status: { type: integer, example: 404 }
        detail: { type: string, example: plan 01J... not found }
        instance: { type: string }
    Workstream:
      type: object
      required: [id, title, status, priority, created_at, updated_at]
      properties:
        id: { type: string }
        title: { type: string }
        status: { $ref: '#/components/schemas/WorkstreamStatus' }
        priority: { type: integer }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    WorkstreamStatus:
      type: string
      enum: [active, paused, done, archived]
    WorkstreamInput:
      type: object
      required: [title]
      properties:
        title: { type: string }
        status: { $ref: '#/components/schemas/WorkstreamStatus' }
        priority: { type: integer }
...
    WorkStatus:
      type: string
      description: Stored step/task status; `ready` is only ever the derived
        boolean flag, never a status value.
      enum: [pending, in_progress, done, failed, blocked, interrupted, cancelled]
...
    AddedBy:
      type: string
      enum: [planner, orchestrator, user]
...
    Outcome:
      type: object
      required: [result]
      properties:
        result: { type: string }
        evidence:
          type: array
          description: Required non-empty when finishing with `done`.
          items: { type: string }
    Task:
      type: object
      required: [id, workstream_id, requested_by, title, description,
        acceptance_criteria, scope, status, attempt, added_by, version,
        created_at, updated_at]
      properties:
        id: { type: string }
        workstream_id: { type: string }
        requested_by: { type: string }
        title: { type: string }
        description: { type: string }
        acceptance_criteria: { type: string }
        scope: { type: object, additionalProperties: true }
        status: { $ref: '#/components/schemas/WorkStatus' }
        assignee_agent_id: { type: string, nullable: true }
        outcome:
          allOf:
            - $ref: '#/components/schemas/Outcome'
          nullable: true
        attempt: { type: integer }
        added_by: { $ref: '#/components/schemas/AddedBy' }
        version: { type: integer }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    TaskInput:
      type: object
      required: [workstream_id, requested_by, title, added_by]
      properties:
        workstream_id: { type: string }
        requested_by: { type: string }
        title: { type: string }
        description: { type: string }
        acceptance_criteria: { type: string }
        scope: { type: object, additionalProperties: true }
        added_by: { $ref: '#/components/schemas/AddedBy' }
...
    Page:
      type: object
      description: Standard list envelope; `next_cursor` is null on the last page.
      required: [items]
      properties:
        next_cursor: { type: string, nullable: true }
...
    TaskPage:
      allOf:
        - $ref: '#/components/schemas/Page'
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items: { $ref: '#/components/schemas/Task' }
...